Review feedback
- Remove unused RXJS import
- Move Monitor status and icon maps to a separate file
